+++ raytracer/CHANGELOG.md
@@ -13,15 +13,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - A `primary_axial_color` method to `ParaxialView` for computing the axial primary color aberration of a lens.
 - An `axes` method on `SequentialModel` to return the set of axes that the system is modeled over.
 - `RayBundle`, `TraceResultsCollection` were added as part of refactoring the `ray_trace_3d_view`.
+- A f = +100 mm biconvex lens example with an object at a finite distance.
 
 ### Changed
 
 - `RefractiveIndexSpec` is now a trait which supports getting refractive index data from any generic materials database.
 - `ray_trace_3d_view` now returns a `TraceResultsCollection` of modified `TraceResults` structs. This allows for better access to a given set of values for (field_id, wavelength_id, Axis). `Ray` was also modified and now contains only position and direction information.
+- `Ray::fan` was removed from the public API and is now public only from within the crate.
+- Renamed the `ObjectHeight` field spec to `PointSource` to reflect that it is a point source of light. Its inputs are now the (x, y) position of the point source within the object plane.
+- Renamed the `ChiefAndMarginalRays` `PupilSampling` variant to `TangentialRayFan`.
 
 ### Fixed
 
 - Fixed an import error in the `n` macro.
+- `Vec3::normalize` is now a no-op when a zero-length vector is provided. This avoids the introduction of NaNs into calculations.
+- Radius vs. diameter mixup when computing paralel ray bundles on a square grid on the entrance pupil in the `ray_trace_3d_view`.
+- Chief ray calculations for finite object fields by `ParaxialSubview` now return the correct sign for the initial paraxial angle.
 
 ## [0.2.0] 2024-12-09